feat: problem 58
