+++ src/lib.rs
@@ -64,3 +64,4 @@ pub mod problem_59;
 pub mod problem_60;
 pub mod problem_61;
 pub mod problem_62;
+pub mod problem_63;
